Fix SSR (#971) (#1456)
- Add env variable to flag manual style resets
- Export minimum and recommended style resets
- Export function to apply resets manually with ability to provide
custom applying funcion
